Add files via upload
This script is designed to collect face data for use in face recognition applications. It prompts the user to enter a name and then captures images from the webcam, detects faces in the images, and saves the cropped face images along with the corresponding name. The collected data can be used to train a face recognition model.